[test] Move useful helpers into StdlibUnicodeUnittest
